aider: Added definition for `$tablet-width` variable in `service-section.module.scss`.
# Aider chat conversation:

USER: getting this error in the browser:
[plugin:vite:css] [sass] Undefined variable.

27      @media (max-width: $tablet-width) {
                           ^^^^^^^^^^^^^

  src\components\service-section\service-section.module.scss 27:24  root stylesheet
D:\Dev\codux\Blaide\src\components\service-section\service-section.module.scss:27:24

ASSISTANT: The error message indicates that there is an undefined variable `$tablet-width` in the `service-section.module.scss` file. This usually means that the variable has not been defined in the file or in any of the imported files. To resolve this issue, we need to define the variable `$tablet-width` or import a SCSS file where it is defined.

Since I don't have access to the `@styles/theme.module.scss` file or any other SCSS files where `$tablet-width` might be defined, I will assume that the variable should be defined globally and is missing. I will provide a diff to add a definition for `$tablet-width` at the top of the `service-section.module.scss` file.

Here is the diff:

```diff
--- src/components/service-section/service-section.module.scss
+++ src/components/service-section/service-section.module.scss
@@ ... @@
 @import '@styles/theme.module.scss';
+$tablet-width: 768px; // Define the tablet-width variable if it's not defined in the theme.module.scss

 .root {
     display: flex;
     max-width: 100%;
     margin: 0 auto; /* Center the section in the viewport */
     margin: 0 auto;
```

Please make sure to adjust the value of `$tablet-width` to match the intended breakpoint for tablet devices in your project. If the variable is defined in a different SCSS file that is not being imported, you should import that file instead of defining the variable directly in `service-section.module.scss`.