aider: Update header.module.scss to use $button-text instead of $button-font to fix undefined variable error.
# Aider chat conversation:

USER: thanks. iam seeing this error in the browser after your last commit :
[plugin:vite:css] [sass] Undefined variable.

10      font: $button-font;
              ^^^^^^^^^^^^

  src\components\header\header.module.scss 10:11  root stylesheet

ASSISTANT: It seems like the variable `$button-font` is not defined in your `theme.module.scss` or any other imported SCSS file. To fix this, we need to define `$button-font` in `theme.module.scss` or update the `header.module.scss` to use an existing variable for the font.

Since I don't have the exact font details that `$button-font` should correspond to, I'll assume that it should be the same as `$button-text` which is already defined in your `theme.module.scss`. Here's the diff to update `header.module.scss` to use `$button-text`:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ ... @@
 .menu {
     display: flex;
     column-gap: 2rem;
-    font: $button-font;
+    font: $button-text;
     color: $black;
 }
```

Apply this change to your `header.module.scss` file, and it should resolve the undefined variable error you're encountering.