aider: Replace `$button-font` with `$button-text` in `header.module.scss` to resolve undefined variable error.
# Aider chat conversation:

USER: still getting this error in the browser:
[plugin:vite:css] [sass] Undefined variable.

10      font: $button-font;
              ^^^^^^^^^^^^

  src\components\header\header.module.scss 10:11  root stylesheet

ASSISTANT: It appears that the `$button-font` variable is still being referenced in the `header.module.scss` file, but it is not defined in the `theme.module.scss` or any other imported SCSS file. To resolve this issue, we need to either define `$button-font` in the `theme.module.scss` file or replace the `$button-font` variable with an existing variable in the `header.module.scss` file.

Since the `theme.module.scss` file contains a variable `$button-text` that seems to be intended for button font styling, I will provide a diff to replace `$button-font` with `$button-text` in the `header.module.scss` file:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ ... @@
 .root {
     display: flex;
     flex-direction: row;
     padding: 0 3em; // Unify padding for left and right
     justify-content: space-between;
     align-items: center;
     height: 4em;
-    font: $button-font;
+    font: $button-text;
     background-color: $white; // Use variable for consistency
     top: 0;
     left: 0;
     width: 100%;
     z-index: 1000; /* Ensure the header is above other content */
```

Please apply this change to your `header.module.scss` file, and it should resolve the undefined variable error.