aider: Corrected reference to `$button-font` variable in `header.module.scss` to `$button-text`.
# Aider chat conversation:

USER: now this :
[plugin:vite:css] [sass] Undefined variable.

10      font: $button-font;
              ^^^^^^^^^^^^

  src\components\header\header.module.scss 10:11  root stylesheet

ASSISTANT: It seems that the error persists because the variable `$button-font` is still being referenced in the `header.module.scss` file. Since the variable `$button-text` is defined in the `theme.module.scss` and was intended to be used, I will provide a diff to correct the reference in the `header.module.scss` file.

Here's the diff to correct the variable reference:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ -7,7 +7,7 @@
     justify-content: space-between;
     align-items: center;
     height: 4em;
-    font: $button-font;
+    font: $button-text;
     background-color: $white; // Use variable for consistency
     top: 0;
     left: 0;
```

Please apply this change to your `header.module.scss` file, and it should resolve the undefined variable error.