aider: Corrected import path for theme module in header module.
# Aider chat conversation:

USER: i still get this :
[plugin:vite:css] [sass] Undefined variable.

18      @media only screen and (max-width: $mobile-width) {
                                           ^^^^^^^^^^^^^

  src\components\header\header.module.scss 18:40  root stylesheet
D:\Dev\codux\Blaide\src\components\header\header.module.scss:18:40

ASSISTANT: The error message indicates that the variable `$mobile-width` is not defined or not accessible within the `header.module.scss` file. This variable is supposed to be defined in the `theme.module.scss` file, which should be imported at the top of your `header.module.scss` file.

Let's ensure that the variable `$mobile-width` is correctly defined in the `theme.module.scss` file and that this file is correctly imported in `header.module.scss`. I will provide the necessary diff to correct the import statement in `header.module.scss`:

```diff
--- src/components/header/header.module.scss
+++ src/components/header/header.module.scss
@@ -1,4 +1,4 @@
-@import '@styles/theme.module.scss';
+@import '../../styles/theme.module.scss';

 .root {
     display: flex;
```

This change corrects the import path to the `theme.module.scss` file, assuming that the `styles` directory is located at the root of the `src` directory. Please apply this change, and it should resolve the undefined variable error for `$mobile-width`.